Replace unsigned int declarations with size_t and improve clearing of the forceField link list.

+++ modules/SofaGeneralAnimationLoop/MechanicalMatrixMapper.h
@@ -170,7 +170,7 @@ protected:
 
     Data<helper::vector<std::string>> d_forceFieldList;
 
-    unsigned int m_nbInteractionForceFields;
+    size_t m_nbInteractionForceFields;
 
     MechanicalMatrixMapper() ;
